gitbak

+++ Reference_Implementation/include/LESS.h
@@ -31,21 +31,20 @@
 #include <stddef.h>
 
 
-/* Public key: the first gen. matrix is shrunk to just a seed, all the
- * others are stored in RREF form  */
+/// Public key: the first gen. matrix is shrunk to just a seed, all the
+/// others are stored in RREF form
 typedef struct __attribute__((packed)) {
    unsigned char G_0_seed[SEED_LENGTH_BYTES];
    uint8_t SF_G [NUM_KEYPAIRS-1][RREF_MAT_PACKEDBYTES];
 } pubkey_t;
 
-/* Private key: it contains both a single seed generating all private *
- * (inverse) monomials and the seed to generate the public code */
+/// Private key: it contains  a single seed generating all private (inverse) monomials
 typedef struct __attribute__((packed)) {
-   /*the private key is compressible down to a single seed*/
    unsigned char compressed_sk[PRIVATE_KEY_SEED_LENGTH_BYTES];
 } prikey_t;
 
-
+/// signature: containing the commitment digest and hash. As well as the public 
+/// information sets of the canonical forms, as well as the seed tree
 typedef struct __attribute__((packed)) sig_t {
     uint8_t digest[HASH_DIGEST_LENGTH];
     uint8_t salt[HASH_DIGEST_LENGTH];
@@ -53,18 +52,32 @@ typedef struct __attribute__((packed)) sig_t {
     uint8_t seed_storage[SEED_TREE_MAX_PUBLISHED_BYTES];
 } sign_t;
 
-/* keygen cannot fail */
+/// keygen cannot fail
+/// \param SK[out]: pointer to an uninitialized secret key data structure
+/// \param PK[out]: pointer to an uninitialized public key data structure
 void LESS_keygen(prikey_t *SK,
                  pubkey_t *PK);
 
-/* sign cannot fail, but it returns the number of opened seeds */
+/// sign cannot fail, but it returns the number of opened seeds
+/// \param SK[in]: pointer to an initialized secret key
+/// \param m[in]: pointer to the message to sign
+/// \param mlen[in]: length of the message
+/// \param sig[out]: pointer to a `sign_t` structure.
+/// \returns number of opened seeds
 size_t LESS_sign(const prikey_t *SK,
-               const char *const m,
-               const uint64_t mlen,
-               sign_t *sig);
+                 const char *const m,
+                 const size_t mlen,
+                 sign_t *sig);
 
-/* verify returns 1 if signature is ok, 0 otherwise */
+/// verify returns 1 if signature is ok, 0 otherwise
+/// \param PK[in]: pointer to an initialized public key
+/// \param m[in]: pointer to the message to sign
+/// \param mlen[in]: length of the message
+/// \param sig[in]: pointer to a `sign_t` structure, containing a hopefully  
+///     correct signature
+/// \return: 1 on success, 
+///          0 on failure
 int LESS_verify(const pubkey_t *const PK,
                 const char *const m,
-                const uint64_t mlen,
+                const size_t mlen,
                 const sign_t *const sig);
